# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PaymentNotificationMethod = typeof PaymentNotificationMethod[keyof typeof PaymentNotificationMethod];


export const PaymentNotificationMethod = {
  card: 'card',
  wire: 'wire',
} as const;

export interface PaymentNotification {
  method: PaymentNotificationMethod;
}

export interface RemittanceSubmission {
  objectPath: string;
  fileName: string;
  note?: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface ErrorEnvelope {
  error: string;
}