# lib/api-zod/src/generated/types/remittanceSubmission.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RemittanceSubmission {
  objectPath: string;
  fileName: string;
  note?: string;
}